Fix : 정지 해제 시 등록 정보 업데이트 수정

+++ src/main/java/com/soocompany/wodify/registration_info/domain/RegistrationInfo.java
@@ -45,7 +45,7 @@ public class RegistrationInfo extends BaseEntity {
                 .build();
     }
 
-    public void updateEndDate(int holdingPeriod) {
+    public void updateEndDate(long holdingPeriod) {
         this.endDate = this.endDate.plusDays(holdingPeriod);
     }
 }
